linux installer

+++ toolchain-setup.py
@@ -20,20 +20,12 @@ print(text.blue + "Open Orbis Toolchain Setup\n" + text.reset)
 
 # Get installation parameters
 install_type = int(input(
-                   "Set installation type (WSL = 1 | WINDOWS =2 | LINUX = 3): "))
+                   "Set installation type (WSL = 1 | WINDOWS = 2 | LINUX = 3): "))
 
 
 if install_type < 1 or install_type > 3:
-    print("Installation not supported")
+    print("Installation type not supported")
     exit(-1)
-
-
-if install_type == install_types.WSL:
-    None
-elif install_type == install_types.WINDOWS:
-    None
-elif install_type == install_types.LINUX:
-    None
 
 print(text.clear)
 if environment_name in os.environ:
@@ -64,6 +56,7 @@ elif install_type == install_types.LINUX:
     toolchain_url = "https://github.com/OpenOrbis/OpenOrbis-PS4-Toolchain/releases/download/v0.5.1/v0.5.1.tar.gz"
     filename = wget.download(toolchain_url, out=output_dir)
 
+
 # Install Toolchain Windows
 toolchain_dir = ""
 if install_type == install_types.WSL or install_type == install_types.WINDOWS:
@@ -84,9 +77,21 @@ if install_type == install_types.WSL or install_type == install_types.WINDOWS:
 
 # Install Toolchain Linux
 elif install_type == install_types.LINUX:
-    install_dir = input("Select Toolchain installation path (default:)")
-    # TODO extract tar.gz to installation dir
-    None
+    home_dir = os.environ.get("HOME")
+    toolchain_dir = os.path.join(home_dir, "OpenOrbis/PS4Toolchain/")
+    print(text.clear)
+    conti = input("Toolchain will be installed in '" +
+                  toolchain_dir+"'. Continue? (Y/N): ")
+    while conti != "Y" and conti != "N":
+        conti = input("Toolchain will be installed in '" +
+                      toolchain_dir+"'. Continue? (Y/N): ")
+    if conti == "Y":
+        file_path = os.path.join(output_dir, filename)
+        file = tarfile.open(file_path)
+        file.extractall(toolchain_dir)
+        file.close()
+    else:
+        toolchain_dir = ""
 
 # Clean installation files
 print("Cleaning installation files")
@@ -94,40 +99,26 @@ installer_path = os.path.join(output_dir, filename)
 if exists(installer_path):
     os.remove(installer_path)
 
-# Set Environment variable OO_PS4_TOOLCHAIN
-set_env = ""
-while set_env != "Y" and set_env != "N":
-    set_env = input("Set environment variable 'OO_PS4_TOOLCHAIN' ? (Y/N): ")
-
-# Set env WINDOWS
-if set_env == "Y":
-    if install_type == install_types.WINDOWS:
-        # TODO set env windows
-        None
-    elif install_type == install_types.LINUX:
-        # TODO set env linux
-        None
-    elif install_type == install_types.WSL:
-        print("Cannot set env for wsl")
-
 # Download create-gp4
 install_create_gp4 = input("Add 'create-gp4' to Toolchain? (Y/N): ")
 while install_create_gp4 != "Y" and install_create_gp4 != "N":
-    install_create_gp4 = input(
-        "Set environment variable 'OO_PS4_TOOLCHAIN' ? (Y/N): ")
+    install_create_gp4 = input("Add 'create-gp4' to Toolchain? (Y/N): ")
+if toolchain_dir != "":
+    if install_create_gp4 == "Y":
+        # Download create-gp4.tar.gz
+        create_gp4_url = "https://github.com/OpenOrbis/create-gp4/releases/download/v1.0/binaries.tar.gz"
+        filename = wget.download(create_gp4_url, out=output_dir)
+        file_path = os.path.join(output_dir, filename)
 
-if install_create_gp4 == "Y":
-    # Download create-gp4.tar.gz
-    create_gp4_url = "https://github.com/OpenOrbis/create-gp4/releases/download/v1.0/binaries.tar.gz"
-    filename = wget.download(create_gp4_url, out=output_dir)
-    file_path = os.path.join(output_dir, filename)
+        # Extract to toolchain dir
+        file = tarfile.open(file_path)
+        file.extractall(toolchain_dir)
+        file.close()
+        # Clean files
+        if exists(file_path):
+            os.remove(file_path)
+else:
+    print("Toolchain directory not found")
+    exit(-1)
 
-    # Extract to toolchain dir
-    file = tarfile.open(file_path)
-    file.extractall(toolchain_dir)
-    file.close()
-    # Clean files
-    if exists(file_path):
-        os.remove(file_path)
-
-print(text.blue + "OpenOrbis Toolchain configured succesfully" + text.reset)
+print(text.blue + "\nOpenOrbis Toolchain configured succesfully" + text.reset)
